+++ graphql-java-server-runtime/src/test/java/com/graphql_java_generator/server/util/test_classes/AnInterface.java
@@ -1,0 +1,14 @@
+package com.graphql_java_generator.server.util.test_classes;
+
+import com.graphql_java_generator.annotation.GraphQLInterfaceType;
+
+/**
+ * A class to test the {@link com.graphql_java_generator.server.util.GraphQLServerUtils#classNameExtractor(Class)}
+ * method
+ * 
+ * @author etienne-sf
+ */
+@GraphQLInterfaceType("TheInterfaceName")
+public class AnInterface {
+
+}