+++ graphql-java-server-runtime/src/test/java/com/graphql_java_generator/server/util/test_classes/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains the test classes for the
+ * {@link com.graphql_java_generator.server.util.GraphQLServerUtils#classNameExtractor(Class)} method
+ * 
+ */
+package com.graphql_java_generator.server.util.test_classes;