+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateMyQueryTypeImpl.java
@@ -7,15 +7,25 @@ import java.util.UUID;
 
 import javax.annotation.Resource;
 
-import org.allGraphQLCases.server.*;
-import org.allGraphQLCases.server.STP_AllFieldCases_STS;
-import org.allGraphQLCases.server.SUP_AnyCharacter_SUS;
+import org.allGraphQLCases.server.SEP_EnumWithReservedJavaKeywordAsValues_SES;
+import org.allGraphQLCases.server.SEP_Episode_SES;
+import org.allGraphQLCases.server.SEP_extends_SES;
+import org.allGraphQLCases.server.SINP_AllFieldCasesInput_SINS;
 import org.allGraphQLCases.server.SINP_CharacterInput_SINS;
 import org.allGraphQLCases.server.SINP_DroidInput_SINS;
-import org.allGraphQLCases.server.SEP_EnumWithReservedJavaKeywordAsValues_SES;
-import org.allGraphQLCases.server.STP_Foo140_STS;
 import org.allGraphQLCases.server.SINP_HumanInput_SINS;
+import org.allGraphQLCases.server.SIP_CharacterConnection_SIS;
+import org.allGraphQLCases.server.SIP_Character_SIS;
+import org.allGraphQLCases.server.SIP_Client_SIS;
+import org.allGraphQLCases.server.STP_AllFieldCases_STS;
+import org.allGraphQLCases.server.STP_Droid_STS;
+import org.allGraphQLCases.server.STP_Foo140_STS;
+import org.allGraphQLCases.server.STP_HumanConnection_STS;
+import org.allGraphQLCases.server.STP_Human_STS;
 import org.allGraphQLCases.server.STP_MyQueryType_STS;
+import org.allGraphQLCases.server.STP_break_STS;
+import org.allGraphQLCases.server.SUP_AnyCharacter_SUS;
+import org.allGraphQLCases.server.extensions.GraphQlException;
 import org.allGraphQLCases.server.util.DataFetchersDelegateMyQueryType;
 import org.springframework.stereotype.Component;
 
@@ -48,7 +58,8 @@ public class DataFetchersDelegateMyQueryTypeImpl implements DataFetchersDelegate
 	}
 
 	@Override
-	public SIP_Character_SIS withOneOptionalParam(DataFetchingEnvironment dataFetchingEnvironment, SINP_CharacterInput_SINS character) {
+	public SIP_Character_SIS withOneOptionalParam(DataFetchingEnvironment dataFetchingEnvironment,
+			SINP_CharacterInput_SINS character) {
 		if (character == null) {
 			return generator.generateInstance(STP_Human_STS.class);
 		} else {
@@ -67,14 +78,16 @@ public class DataFetchersDelegateMyQueryTypeImpl implements DataFetchersDelegate
 	}
 
 	@Override
-	public SIP_Character_SIS withOneMandatoryParam(DataFetchingEnvironment dataFetchingEnvironment, SINP_CharacterInput_SINS character) {
+	public SIP_Character_SIS withOneMandatoryParam(DataFetchingEnvironment dataFetchingEnvironment,
+			SINP_CharacterInput_SINS character) {
 		SIP_Character_SIS c = mapper.map(character, getClassFromName(SIP_Character_SIS.class, character.getType()));
 		c.setId(UUID.randomUUID());
 		return c;
 	}
 
 	@Override
-	public SIP_Character_SIS withEnum(DataFetchingEnvironment dataFetchingEnvironment, SEP_Episode_SES SEP_Episode_SES) {
+	public SIP_Character_SIS withEnum(DataFetchingEnvironment dataFetchingEnvironment,
+			SEP_Episode_SES SEP_Episode_SES) {
 		SIP_Character_SIS c = generator.generateInstance(STP_Droid_STS.class);
 
 		// The SEP_Episode_SES list (appearsIn) will be filled by another call (the graphql manages the joins).
@@ -85,7 +98,8 @@ public class DataFetchersDelegateMyQueryTypeImpl implements DataFetchersDelegate
 	}
 
 	@Override
-	public STP_AllFieldCases_STS withListOfList(DataFetchingEnvironment dataFetchingEnvironment, List<List<Double>> matrix) {
+	public STP_AllFieldCases_STS withListOfList(DataFetchingEnvironment dataFetchingEnvironment,
+			List<List<Double>> matrix) {
 		STP_AllFieldCases_STS ret = new STP_AllFieldCases_STS();
 		ret.setMatrix(matrix);
 		return ret;
@@ -124,11 +138,12 @@ public class DataFetchersDelegateMyQueryTypeImpl implements DataFetchersDelegate
 	@Override
 	public SIP_Character_SIS error(DataFetchingEnvironment dataFetchingEnvironment, String errorLabel) {
 		// This method is here only to test the error behavior.
-		throw new RuntimeException("This is an error: " + errorLabel);
+		throw new GraphQlException("This is an error: " + errorLabel);
 	}
 
 	@Override
-	public STP_AllFieldCases_STS allFieldCases(DataFetchingEnvironment dataFetchingEnvironment, SINP_AllFieldCasesInput_SINS input) {
+	public STP_AllFieldCases_STS allFieldCases(DataFetchingEnvironment dataFetchingEnvironment,
+			SINP_AllFieldCasesInput_SINS input) {
 		STP_AllFieldCases_STS ret;
 		if (input != null) {
 			ret = mapper.map(input, STP_AllFieldCases_STS.class);
@@ -237,8 +252,9 @@ public class DataFetchersDelegateMyQueryTypeImpl implements DataFetchersDelegate
 	}
 
 	@Override
-	public List<SUP_AnyCharacter_SUS> unionTest(DataFetchingEnvironment dataFetchingEnvironment, SINP_HumanInput_SINS human1,
-			SINP_HumanInput_SINS human2, SINP_DroidInput_SINS droid1, SINP_DroidInput_SINS droid2) {
+	public List<SUP_AnyCharacter_SUS> unionTest(DataFetchingEnvironment dataFetchingEnvironment,
+			SINP_HumanInput_SINS human1, SINP_HumanInput_SINS human2, SINP_DroidInput_SINS droid1,
+			SINP_DroidInput_SINS droid2) {
 		List<SUP_AnyCharacter_SUS> ret = new ArrayList<>();
 
 		if (human1 != null) {
@@ -277,8 +293,8 @@ public class DataFetchersDelegateMyQueryTypeImpl implements DataFetchersDelegate
 	}
 
 	@Override
-	public STP_Droid_STS withTwoMandatoryParamDefaultVal(DataFetchingEnvironment dataFetchingEnvironment, SINP_DroidInput_SINS theHero,
-			Integer num) {
+	public STP_Droid_STS withTwoMandatoryParamDefaultVal(DataFetchingEnvironment dataFetchingEnvironment,
+			SINP_DroidInput_SINS theHero, Integer num) {
 		// TODO Auto-generated method stub
 		return null;
 	}