+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/oauth2/WebSecurityConfiguration.java
@@ -1,6 +1,5 @@
 package org.allGraphQLCases.server.oauth2;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
@@ -14,25 +13,17 @@ import org.springframework.security.web.server.SecurityWebFilterChain;
 @EnableWebFluxSecurity
 public class WebSecurityConfiguration {
 
-	@Value("${spring.security.oauth2.resource.introspection-uri}")
-	String introspectionUri;
-
-	@Value("${spring.security.oauth2.client.client-id}")
-	String clientId;
-
-	@Value("${spring.security.oauth2.client.client-secret}")
-	String clientSecret;
-
 	@Bean
 	SecurityWebFilterChain configure(ServerHttpSecurity http) throws Exception {
-		return http//
-					// Disabling CORS and CSRF makes POST on the graphql URL work properly. Double-check that before
-					// entering in production
+		http//
+			// Disabling CORS and CSRF makes POST on the graphql URL work properly. Double-check that before
+			// entering in production
 				.cors().and().csrf().disable()//
-				.authorizeExchange().pathMatchers("/my/updated/graphql/path").authenticated()//
-				.and().oauth2Login() // oauth2Client()
-				// .access("hasRole('ROLE_CLIENT')")//
-				.and().build();
+				.authorizeExchange().pathMatchers("/my/updated/graphql/path").authenticated()
+				// .hasAuthority("ROLE_CLIENT")//
+				.and()//
+				.oauth2ResourceServer().jwt();
+		return http.build();
 	}
 
 }